% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSE-package.R
\docType{package}
\name{ABSE-package}
\alias{ABSE}
\alias{ABSE-package}
\title{ABSE: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
