% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbe-package.R
\docType{package}
\name{bbe-package}
\alias{bbe}
\alias{bbe-package}
\title{bbe: Basin-Based Evaluation (BBE) of Multiobjective Optimization Algorithm Runs}
\description{
This package provides a customizable method for the basin-based evaluation of multiobjective optimization algorithm runs focusing on multimodality. The main idea is that the decision space is partitioned in basins which are evaluated separately with a user-defined function.
}
\author{
\strong{Maintainer}: Jonathan Heins \email{jonathan.heins@tu-dresden.de} (\href{https://orcid.org/0000-0002-3571-667X}{ORCID})

}
\keyword{internal}
