% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{filterByBasin}
\alias{filterByBasin}
\title{Group a set of points by the label of the region in the rasterized decision space htey lie in.}
\usage{
filterByBasin(solutions, basinLabels, boudaries, nBasins, gridSize, nDim)
}
\arguments{
\item{solutions}{A data frame with one column per dimension.}

\item{basinLabels}{A vector containing the labels of all regions as retrieved from getBasinLabels.}

\item{nBasins}{The number of unique labels in \code{\link{basinLabels}}.}

\item{gridSize}{The side length of the grid.}

\item{nDim}{The number of dimensions in the decision space.}

\item{boundaries}{A vector containing the minimal and maximal possible values per dimension.}
}
\description{
Group a set of points by the label of the region in the rasterized decision space htey lie in.
}
