% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBasinLabels}
\alias{getBasinLabels}
\title{Identify for every point in the rasterised decision space the corresponding basin it belongs to.}
\usage{
getBasinLabels(efficientSets, gradients, gridSize, nDim)
}
\arguments{
\item{efficientSets}{A list containing for every efficient set the indices of all efficient points.}

\item{gradients}{A vector containing the accumulated gradients for every point in the grid.}

\item{gridSize}{The side length of the grid.}

\item{nDim}{The number of dimensions in the decision space.}
}
\description{
Identify for every point in the rasterised decision space the corresponding basin it belongs to.
}
