% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getEfficientSets}
\alias{getEfficientSets}
\title{For a vector of indices with efficient points within a grid identify efficient sets.}
\usage{
getEfficientSets(
  efficientPoints,
  gridSize,
  nDim,
  domSort = FALSE,
  rank = as.numeric(c()),
  nRank = 0L,
  joinFronts = FALSE
)
}
\arguments{
\item{efficientPoints}{The vector containing the indices of all efficient points.}

\item{gridSize}{The side length of the grid.}

\item{nDim}{The number of dimensions in the decision space.}
}
\description{
For a vector of indices with efficient points within a grid identify efficient sets.
}
